change way to get directory and read java file

+++ ChangePath/Form1.cs
@@ -24,14 +24,17 @@ namespace ChangePath
         {
 
             string workingDirectory = Environment.CurrentDirectory;
-            string projectDirectory = Directory.GetParent(workingDirectory).Parent.FullName;
-            string fullPath = Path.Combine(projectDirectory, "AutoSource",  "Example.java");
+            //string projectDirectory = Directory.GetParent(workingDirectory).Parent.FullName;
+            //string fullPath = Path.Combine(projectDirectory, "AutoSource",  "Example.java");
+            string peerDirectory = Directory.GetDirectoryRoot(workingDirectory);
+            string fullPath = Path.Combine(peerDirectory, @"EZMS_Automation\src\TestLab", "Share_query.java");
 
             // create new .java file
             CreateFile(fullPath, txtPath.Text);
 
             // run bat file
-            fullPath = Path.Combine(projectDirectory, "AutoSource", "run.bat");
+            //fullPath = Path.Combine(projectDirectory, "AutoSource", "run.bat");
+            fullPath = Path.Combine(peerDirectory, @"EZMS_Automation", "Download_Attachment_VSTS.bat");
             int returnCode = ExecuteCommand(fullPath);
             if (returnCode > -1)
                 MessageBox.Show("Run successfully!!!!");
@@ -43,21 +46,27 @@ namespace ChangePath
         {
             try
             {
+                List<string> lineToWrite = new List<string>();
+                string line = null;
+                using (StreamReader reader = new StreamReader(fullPath))
+                {
+                    while (!reader.EndOfStream)
+                    {
+                        line = reader.ReadLine().ToString();
+                        if (line.Contains("Common_Action_Chrome.DonwloadAttachmentVSTS("))
+                            line = "\t\tCommon_Action_Chrome.DonwloadAttachmentVSTS(\"" + newPath.Replace(@"\", @"\\") + "\", \"null\");";
+                        lineToWrite.Add(line);
+                    }
+                }
+                if (lineToWrite == null)
+                    throw new InvalidDataException("Line does not exist in " + fullPath);
+
                 if (File.Exists(fullPath))
                 {
                     File.Delete(fullPath);
                 }
-                // Create a new file     
-                using (StreamWriter sw = File.CreateText(fullPath))
-                {
-                    sw.WriteLine("package TestLab;\r\n");
-                    sw.WriteLine("import java.util.Arrays;\r\n");
-                    //....
-                    sw.WriteLine("Common_Action_Chrome.DonwloadAttachmentVSTS(\"" + newPath.Replace(@"\", @"\\") + "\", \"null\");");
-                    //sw.WriteLine("Add one more line ");
-                    //sw.WriteLine("Done! ");
-
-                }
+                //Create a new file
+                File.WriteAllLines(fullPath, lineToWrite);
             }
             catch (Exception Ex)
             {
